move renderDataSourceConfiguration to helpers.ts
